+++ src/project/types.ts
@@ -1,0 +1,6 @@
+import type { BackendIdentifier } from "@aws-amplify/plugin-types";
+
+export interface AmplifyProject {
+  getStackName(): string | undefined;
+  getBackendIdentifier(): BackendIdentifier | undefined;
+}
